tidy DFG code structure and fix most warnings;
also introduces default-multiple token files for multiple file parts

+++ src/io/FileInputOutput.hpp
@@ -9,7 +9,7 @@
 #include "utils/Types.hpp"
 
 namespace combigrid {
-void writeConcatenatedFileRootOnly(const char* data, int sizeOfData, const std::string& path,
+void writeConcatenatedFileRootOnly(const char* data, size_t sizeOfData, const std::string& path,
                                    MPI_Comm comm, bool replaceExistingFile = false);
 
 bool getFileExistsRootOnly(const std::string& fileName, CommunicatorType comm, RankType myRank,
